Bug reporting via GH forms
## Uses Github's issue forms feature

Documentation: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/syntax-for-issue-forms

Announcement: https://github.blog/changelog/2021-06-23-issues-forms-beta-for-public-repositories/

Example of it in use:
[Feature Flags for Laravel new Issue Page](https://github.com/ylsideas/feature-flags/issues/new/choose)
[Feature Flags for Laravel Bug Report Form](https://github.com/ylsideas/feature-flags/issues/new?assignees=peterfox&labels=bug&template=bug.yml&title=%5BBug%5D%3A+)

---

I've been using this technique for a few months and it's a lot better in guiding users of a package to provide the right information. The form will ask for info like what happened, how to reproduce the bug, the package version, PHP version, Laravel version and which operating systems were used. There is also an optional notes field for further thoughts from the end user.

The Issue link can be removed from `.github/ISSUE_TEMPLATE/config.yml` because the `.github/ISSUE_TEMPLATE/bug.yml` file will be found and automatically displayed on the new issue page.